requirements.txt
fastapi
asyncpg
SQLAlchemy
uvicorn
python-dotenv
pydantic_settings
orjson
jwt
PyJWT[crypto]
passlib
fastapi-cache2[redis]
pydantic[email]
python-multipart
bcrypt
flake8
